feat #79: refactor integration tests

+++ tests/integration_test.rs
@@ -17,8 +17,8 @@ pub fn enable_logger() {
 
 // Unit test port base: 6xxxx
 // Integration test port base: 3xxxx
-pub(crate) const CLUSTER_PORT_BASE: u64 = 30100;
-pub(crate) const APPEND_ENNTRIES_PORT_BASE: u64 = 30200;
-pub(crate) const ELECTION_PORT_BASE: u64 = 30300;
-pub(crate) const JOIN_CLUSTER_PORT_BASE: u64 = 30400;
-pub(crate) const SNAPSHOT_PORT_BASE: u64 = 30500;
+pub(crate) const CLUSTER_PORT_BASE: u16 = 30100;
+pub(crate) const APPEND_ENNTRIES_PORT_BASE: u16 = 30200;
+pub(crate) const ELECTION_PORT_BASE: u16 = 30300;
+pub(crate) const JOIN_CLUSTER_PORT_BASE: u16 = 30400;
+pub(crate) const SNAPSHOT_PORT_BASE: u16 = 30500;
